Add files via upload
This is a simple helpdesk management system that allows the user to login, register, ask queries and give complaints. The pre-requisites are:- Eclipse IDE for Enterprise Java and Web Developers, Tomcat server 9th version, xamp server. The database name is helpdesk. Create tables in database,namely users, queries and complaints.